fix(web): extract team management actions
